--- a/app/pkg/eval/tracer.go
+++ b/app/pkg/eval/tracer.go
@@ -1,0 +1,10 @@
+package eval
+
+import (
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/trace"
+)
+
+func tracer() trace.Tracer {
+	return otel.Tracer("github.com/jlewi/foyle/app/pkg/eval")
+}